Added, Tag Explorer plugin, UDMF: added multiple tags support. Changed, cosmetic: elapsed time is now displayed after resources loading finishes. Changed, internal: changed Clock.CurrentTime type to long. Fixed, Tag Statistics window: in some cases the map view was not updated after selecting items in Sectors/Linedefs/Things columns. Fixed, cosmetic: Draw Curve mode icon was missing a shadow.

+++ Source/Core/Compilers/Compiler.cs
@@ -100,7 +100,7 @@ namespace CodeImp.DoomBuilder.Compilers
 			if(!isdisposed)
 			{
 				Exception deleteerror;
-				float starttime = Clock.CurrentTime;
+				long starttime = Clock.CurrentTime;
 				
 				do
 				{
